Add javascript.info array exercises & misc files
Added the javascript.info array exercises in the new.js file. There are
a few other files here added for a few other unimportant/very short
exercises that I will not list in detail. There is also the
rock-paper-scissors.js along with the paper.html file that does include
the important parts of the rock-paper-scissors console game logic. This
will be expanded upon in future commits as a complete project. It is so
small for now that I didn't commit earlier.