Update file structure to include js Full-Stack
Needed to move all of the files from foundations into the
jsFullStackJavascript folder. Specifically all of these are now in the
intermediateHtmlCss folder since that's where I am currently in the
curriculum. I will continue to add the appropriate folders from now on
for the other sections in the full-stack js course